Main page dynamic / keep track of current workout
the main page is now dynamically generated, which means that once you
start a workout, the button to start a new one disappears until you end
the current one.

The app now keeps track of the current workout in its database, which
means that you can continue a workout even after you close the app. The
app also automatically ends a workout 2 and half hours after it has been
started.